+++ view/index.view.js
@@ -430,6 +430,16 @@ function getCurrentMoveInterpretation() {
     var tp = getCurrentTreePosition()
     var node = tp[0].nodes[tp[1]]
 
+    // Determine and of main variation
+
+    if (gametree.onMainTrack(tp[0]) && gametree.navigate(tp[0], tp[1], 1)[0] == null) {
+        var rootNode = getRootTree().nodes[0]
+
+        if ('RE' in rootNode && rootNode.RE[0].trim() != '') {
+            return 'Result: ' + rootNode.RE[0]
+        }
+    }
+
     // Determine capture
 
     var ptp = gametree.navigate.apply(null, tp.concat([-1]))